pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.0",
    "feedparser>=6.0.11",
    "pydantic>=2.10.6",
]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
